pyproject.toml
[tool.poetry]
name = "web-crawler"
version = "1.0.0"
description = "Web crawler"
readme = "README.md"
authors = ["Florencio Paucar <florenciopaucar@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.scripts]
main = "src.__main__:main"

[tool.poetry.dependencies]
python            = "^3.10"
aiohttp           = "3.8.3"
PyYAML            = "6.0"
pydantic          = "1.9.0"
Flask             = "2.2.2"
"Flask[async]"    = "2.2.2"

[tool.poetry.group.dev.dependencies]
pytest     = "7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"